+++ java/connector-node/risingwave-connector-service/src/main/java/com/risingwave/connector/ConnectorServiceImpl.java
@@ -16,8 +16,8 @@ package com.risingwave.connector;
 
 import com.risingwave.proto.ConnectorServiceGrpc;
 import com.risingwave.proto.ConnectorServiceProto;
-import com.risingwave.sourcenode.SourceRequestHandler;
-import com.risingwave.sourcenode.SourceValidateHandler;
+import com.risingwave.connector.source.SourceRequestHandler;
+import com.risingwave.connector.source.SourceValidateHandler;
 import io.grpc.stub.StreamObserver;
 
 public class ConnectorServiceImpl extends ConnectorServiceGrpc.ConnectorServiceImplBase {
