+++ java/connector-node/risingwave-connector-service/src/main/java/com/risingwave/connector/source/common/PostgresValidator.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.risingwave.sourcenode.common;
+package com.risingwave.connector.source.common;
 
 import com.risingwave.proto.ConnectorServiceProto;
 import com.risingwave.proto.Data;
